+++ src/routes/(protected)/chat/[conversationId]/message/Message.svelte
@@ -15,6 +15,8 @@
 	import TextMessage from "./TextMessage.svelte";
 	import UnsentMessage from "./UnsentMessage.svelte";
 	import UnsupportedMessage from "./UnsupportedMessage.svelte";
+	import AlbumReactionMessage from "./AlbumReactionMessage.svelte";
+	import LocationMessage from "./LocationMessage.svelte";
 
 	let {
 		message,
@@ -174,6 +176,12 @@
 			<ImageMessage message={message.body} />
 		{:else if message.type === "Album" || message.type === "ExpiringAlbum" || message.type === "ExpiringAlbumV2"}
 			<AlbumMessage message={message.body} {isOut} />
+		{:else if message.type === "AlbumContentReaction"}
+			<AlbumReactionMessage message={message.body} />
+		{:else if message.type === "AlbumContentReply"}
+			<AlbumReactionMessage message={message.body} reply={message.body.albumContentReply} />
+		{:else if message.type === "Location"}
+			<LocationMessage message={message.body} />
 		{:else if message.type === "Unsent"}
 			<UnsentMessage />
 		{:else}